+++ src/type/ncontainer.cpp
@@ -13,7 +13,7 @@ namespace ntr
 {
 
 ncontainer::ncontainer(const ntype* iterator_type, const ntype* element_type,
-                       operations* container_ops, uint32_t size, uint32_t align,
+                       operations* container_ops, uint16_t size, uint16_t align,
                        ntype::operations* ops)
     : ntype(ntype::etype::econtainer, size, align, ops), _iterator_type(iterator_type),
       _element_type(element_type), _container_ops(container_ops)
